Only verify TLS to NSX-T Mgr when cert is provided
We acknowledge that we're glossing over a corner case:

When a user has an NSX-T manager with a commercial certificate (e.g.
Let's Encrypt) AND chooses not to supply the certificate THEN the BOSH
vSphere CPI will NOT do TLS validation on the NSX-T Manager (even though
it has the root certificates installed on disk). It's theoretically
possible that this may expose the traffic to MITM (man-in-the-middle)
attack.

We don't believe this corner case exists in real life. We believe almost
all NSX-T Managers have self-signed certificates.

[#176852060]

Co-authored-by: Michael Oleske <moleske@vmware.com>
Co-authored-by: Brian Cunnie <bcunnie@vmware.com>
Co-authored-by: Julian Hjortshoj <hjortshojj@vmware.com>